#37 SMTP issues solved and set up, successful in build and run and testing
Given that the SMTP library from .net is not accessible from
Glimpose.Core due to the MVVMCross structure in place, I placed the SMTP
set up exclusively in Glimpse.Droid, this had to be done as over 10
different plugin failed to work in the core. The testing email used in
this branch is smtp.testing.g@gmail.com, with the password
testingsmtp1234. The email body templates are in the SendMail.cs file in
.Droid.